requirements.txt
streamlit
tensorflow==2.16.1
pillow
numpy
pandas